Bindings

This page documents the low-level bindings to libzmq that were automatically generated. Where possible, the docstrings link to the upstream documentation. Bindings have not been generated for deprecated functions.

Danger

These bindings are unsafe, do not use them unless you know what you're doing.

ZMQ.lib.zmq_atomic_counter_dec — Method

zmq_atomic_counter_dec(counter_)

Upstream documentation.

ZMQ.lib.zmq_atomic_counter_destroy — Method

zmq_atomic_counter_destroy(counter_p_)

Upstream documentation.

ZMQ.lib.zmq_atomic_counter_inc — Method

zmq_atomic_counter_inc(counter_)

Upstream documentation.

ZMQ.lib.zmq_atomic_counter_new — Method

zmq_atomic_counter_new()

Upstream documentation.

ZMQ.lib.zmq_atomic_counter_set — Method

zmq_atomic_counter_set(counter_, value_)

Upstream documentation.

ZMQ.lib.zmq_atomic_counter_value — Method

zmq_atomic_counter_value(counter_)

Upstream documentation.

ZMQ.lib.zmq_bind — Method

zmq_bind(s_, addr_)

Upstream documentation.

ZMQ.lib.zmq_close — Method

zmq_close(s_)

Upstream documentation.

ZMQ.lib.zmq_connect — Method

zmq_connect(s_, addr_)

Upstream documentation.

ZMQ.lib.zmq_ctx_get — Method

zmq_ctx_get(context_, option_)

Upstream documentation.

ZMQ.lib.zmq_ctx_new — Method

zmq_ctx_new()

Upstream documentation.

ZMQ.lib.zmq_ctx_set — Method

zmq_ctx_set(context_, option_, optval_)

Upstream documentation.

ZMQ.lib.zmq_ctx_shutdown — Method

zmq_ctx_shutdown(context_)

Upstream documentation.

ZMQ.lib.zmq_ctx_term — Method

zmq_ctx_term(context_)

Upstream documentation.

ZMQ.lib.zmq_curve_keypair — Method

zmq_curve_keypair(z85_public_key_, z85_secret_key_)

Upstream documentation.

ZMQ.lib.zmq_curve_public — Method

zmq_curve_public(z85_public_key_, z85_secret_key_)

Upstream documentation.

ZMQ.lib.zmq_disconnect — Method

zmq_disconnect(s_, addr_)

Upstream documentation.

ZMQ.lib.zmq_errno — Method

zmq_errno()

Upstream documentation.

ZMQ.lib.zmq_getsockopt — Method

zmq_getsockopt(s_, option_, optval_, optvallen_)

Upstream documentation.

ZMQ.lib.zmq_has — Method

zmq_has(capability_)

Upstream documentation.

ZMQ.lib.zmq_msg_close — Method

zmq_msg_close(msg_)

Upstream documentation.

ZMQ.lib.zmq_msg_copy — Method

zmq_msg_copy(dest_, src_)

Upstream documentation.

ZMQ.lib.zmq_msg_data — Method

zmq_msg_data(msg_)

Upstream documentation.

ZMQ.lib.zmq_msg_get — Method

zmq_msg_get(msg_, property_)

Upstream documentation.

ZMQ.lib.zmq_msg_gets — Method

zmq_msg_gets(msg_, property_)

Upstream documentation.

ZMQ.lib.zmq_msg_init — Method

zmq_msg_init(msg_)

Upstream documentation.

ZMQ.lib.zmq_msg_init_data — Method

zmq_msg_init_data(msg_, data_, size_, ffn_, hint_)

Upstream documentation.

ZMQ.lib.zmq_msg_init_size — Method

zmq_msg_init_size(msg_, size_)

Upstream documentation.

ZMQ.lib.zmq_msg_more — Method

zmq_msg_more(msg_)

Upstream documentation.

ZMQ.lib.zmq_msg_move — Method

zmq_msg_move(dest_, src_)

Upstream documentation.

ZMQ.lib.zmq_msg_recv — Method

zmq_msg_recv(msg_, s_, flags_)

Upstream documentation.

ZMQ.lib.zmq_msg_send — Method

zmq_msg_send(msg_, s_, flags_)

Upstream documentation.

ZMQ.lib.zmq_msg_set — Method

zmq_msg_set(msg_, property_, optval_)

Upstream documentation.

ZMQ.lib.zmq_msg_size — Method

zmq_msg_size(msg_)

Upstream documentation.

ZMQ.lib.zmq_poll — Method

zmq_poll(items_, nitems_, timeout_)

Upstream documentation.

ZMQ.lib.zmq_proxy — Method

zmq_proxy(frontend_, backend_, capture_)

Upstream documentation.

ZMQ.lib.zmq_proxy_steerable — Method

zmq_proxy_steerable(frontend_, backend_, capture_, control_)

Upstream documentation.

ZMQ.lib.zmq_recv — Method

zmq_recv(s_, buf_, len_, flags_)

Upstream documentation.

ZMQ.lib.zmq_send — Method

zmq_send(s_, buf_, len_, flags_)

Upstream documentation.

ZMQ.lib.zmq_send_const — Method

zmq_send_const(s_, buf_, len_, flags_)

Upstream documentation.

ZMQ.lib.zmq_setsockopt — Method

zmq_setsockopt(s_, option_, optval_, optvallen_)

Upstream documentation.

ZMQ.lib.zmq_sleep — Method

zmq_sleep(seconds_)

This is an undocumented function, not part of the formal ZMQ API.

ZMQ.lib.zmq_socket — Method

zmq_socket(arg1, type_)

Upstream documentation.

ZMQ.lib.zmq_socket_monitor — Method

zmq_socket_monitor(s_, addr_, events_)

Upstream documentation.

ZMQ.lib.zmq_stopwatch_intermediate — Method

zmq_stopwatch_intermediate(watch_)

This is an undocumented function, not part of the formal ZMQ API.

ZMQ.lib.zmq_stopwatch_start — Method

zmq_stopwatch_start()

This is an undocumented function, not part of the formal ZMQ API.

ZMQ.lib.zmq_stopwatch_stop — Method

zmq_stopwatch_stop(watch_)

This is an undocumented function, not part of the formal ZMQ API.

ZMQ.lib.zmq_strerror — Method

zmq_strerror(errnum_)

Upstream documentation.

ZMQ.lib.zmq_threadclose — Method

zmq_threadclose(thread_)

This is an undocumented function, not part of the formal ZMQ API.

ZMQ.lib.zmq_threadstart — Method

zmq_threadstart(func_, arg_)

This is an undocumented function, not part of the formal ZMQ API.

ZMQ.lib.zmq_timers_add — Method

zmq_timers_add(timers, interval, handler, arg)

Upstream documentation.

ZMQ.lib.zmq_timers_cancel — Method

zmq_timers_cancel(timers, timer_id)

Upstream documentation.

ZMQ.lib.zmq_timers_destroy — Method

zmq_timers_destroy(timers_p)

Upstream documentation.

ZMQ.lib.zmq_timers_execute — Method

zmq_timers_execute(timers)

Upstream documentation.

ZMQ.lib.zmq_timers_new — Method

zmq_timers_new()

Upstream documentation.

ZMQ.lib.zmq_timers_reset — Method

zmq_timers_reset(timers, timer_id)

Upstream documentation.

ZMQ.lib.zmq_timers_set_interval — Method

zmq_timers_set_interval(timers, timer_id, interval)

Upstream documentation.

ZMQ.lib.zmq_timers_timeout — Method

zmq_timers_timeout(timers)

Upstream documentation.

ZMQ.lib.zmq_unbind — Method

zmq_unbind(s_, addr_)

Upstream documentation.

ZMQ.lib.zmq_version — Method

zmq_version(major_, minor_, patch_)

Upstream documentation.

ZMQ.lib.zmq_z85_decode — Method

zmq_z85_decode(dest_, string_)

Upstream documentation.

ZMQ.lib.zmq_z85_encode — Method

zmq_z85_encode(dest_, data_, size_)

Upstream documentation.